+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [2.0.0] - 2020-03-03
 ### Added
+- New tools to list, clear and update/create oEmbed cached items.
 - Full integration with PerfOps.One suite.
 - Full compatibility with [APCu Manager](https://wordpress.org/plugins/apcu-manager/).
-- New menu (in the left admin bar) for accessing features: "PerfOps Settings".
+- New menus (in the left admin bar) for accessing features: "PerfOps Settings" and "PerfOps Tools".
 ### Changed
 - Cached content may now have a 600K size.
 - Timeout can now be increased up to 1 minute.